app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
joblib==1.4.2
numpy==2.0.1
pandas==2.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlmodel==0.0.21
starlette==0.37.2
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.5
wheel==0.44.0